requirements.txt
blinker==1.9.0
bottle==0.12.25
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
conan==1.51.2
cryptography==44.0.2
Deprecated==1.2.18
et_xmlfile==2.0.0
fasteners==0.19
Flask==3.1.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==6.0.0
MarkupSafe==3.0.2
node-semver==0.6.1
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.3
patch-ng==1.17.4
pluginbase==1.0.1
pycparser==2.22
pyelftools==0.32
PyGithub==2.6.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyserial==3.4
python-dateutil==2.9.0.post0
pytz==2025.1
pywin32==310
PyYAML==6.0
requests==2.32.3
schedule==1.2.2
six==1.16.0
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.1
urllib3==1.26.20
Werkzeug==3.1.3
wrapt==1.17.2
xmltodict==0.14.2
zipp==3.21.0
